refactor AppModel
* add latestLocation to AppModel
* extract Coordinator LocationDelegate
* extract CoreLocation extension
* refactor timer
* don’t use didSet anymore for appState changes (to promptLocationAuth), use Combine and sink
* don’t need appModel.isStationsModalPresented anymore (was to avoid that stationsModel getting closed when timer fired)
